Initial version of Puppetization code
Added a new playbook called puppetize along with a new role with the same name.
The playbook is used to orchestrate the Puppet runs of an OpenStack system while
the puppetize role is used for puppetizing a single host, including setting up
certs on the agent and the puppetmaster sides if needed.

To facilitate single command builds, the new puppetize playbook is a part of
site.yml. This should make it possible to just run site.yml to provision and
configure an OpenStack system from scratch.