Use local user account name for running command on puppetmaster
Instead of trying to run commands as root, run them using the same user
name that is being used to run Ansible. This is not optimal, but
forge-cloud[1-4] don't permit root logins so root won't work. However,
the user name can't be non-root for the other tasks as the freshly
installed hosts to be puppetized don't have any other user accounts on
them.