Removed automatic updating and refactored
Packages are no longer automatically updated. The openstack_common role
became just an install task for Puppet after this, so moved the
installation of Puppet to a more logical place (the puppetize role) so
that Puppet is only installed just in time for running Puppet. Then
removed the openstack_common role as it was now empty. Modified all
those playbooks that called this role to not include it.

Because the openstack_common role was now empty, nothing was actually
being done in the net_servers.yml playbook, so removed that as well.

Added the puppetize role to the hypervisors.yml playbook so that Puppet
would get installed and run there.